Auto-choose skipchain, redirect HTTPs to HTTP
When there is only one skipchain, choose it right away.

Because our production conodes are regrettably talking
websock on HTTP, we need to redirect people who accidentally
connect on https to http, or else websock requests are refused
by the browser. (This is a hack; if this explorer was used in
other contexts, we'd need a better solution.)